chore: fix bucket root behavior
